requirements.txt
appnope==0.1.0
asn1crypto==0.24.0
backcall==0.1.0
bleach==2.1.4
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
cryptography-vectors==2.3.1
Cython==0.29.12
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.2.3
GDAL==2.3.0
gevent==1.4.0
greenlet==0.4.15
grequests==0.3.0
html5lib==1.0.1
idna==2.8
ipykernel==4.10.0
ipython==6.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.1
jedi==0.12.1
Jinja2==2.10
JioMaps==0.1
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.5.0
MarkupSafe==1.0
mistune==0.8.3
mkl-fft==1.0.6
mkl-random==1.0.1
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.6.0
numpy==1.15.2
pandocfilters==1.4.2
parso==0.5.0
pexpect==4.6.0
pickleshare==0.7.4
Pillow==6.1.0
prometheus-client==0.7.1
prompt-toolkit==1.0.15
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.4.2
pyOpenSSL==18.0.0
PySocks==1.6.8
python-dateutil==2.7.3
pyzmq==17.1.2
qtconsole==4.5.1
requests==2.22.0
selenium==3.141.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
terminado==0.8.1
testpath==0.3.1
tornado==5.1.1
tqdm==4.32.2
traitlets==4.3.2
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.1
